[defaults] Add a npm package-lock file
It's apparently good to commit a package lock file to repo's.
"package-lock.json is automatically generated for any operations where
npm modifies either the node_modules tree, or package.json. It describes
the exact tree that was generated, such that subsequent installs are able
to generate identical trees, regardless of intermediate dependency
updates."