[feat] Expose new displayTableWithMainTitle ExtensionHelper method
Sometimes you want the table title to be the main focus in the modal
window that shows the table. The displayTableWithMainTitle() method
in ExtensionHelper has the same format as displayTable() except it
instructs Leopard to take the table's title and override the main
modal's title. This main modal title usually comes from the last
response's answer text.