[feat] Allow you to request location information from the browser
For some conversations it might be desirable to use the user's current
location as a mean of refining a backend query. This new feature will
allow you to define the following new values for an inputType output
parameter.

* locationCityStateZip ( E.g. format returned = Sammamish, Washington 98075 )
* locationZip ( E.g. format returned = 98075 )
* locationLatLong   ( E.g. format returned = lat,log = 40.748736,-73.98486)
* locationJson  ( format returned is url encoded json = see the following
url for the format - https://bit.ly/2Gr6Zyo )

To be able to retrieve any onfo other than "locationLatLong" you will
need to define a license key for https://locationiq.com in .env called:
VUE_APP_LOCATION_IQ_KEY

Teneo will automatically send back the requested information to Teneo.
This will come in as request parameters with the same names:
 (locationCityStateZip/locationZip/locationLatLong/locationJson)
 The parameter values are URL Encoded and you should URL decode them
 retriving the location information. This can be done either in a
 pre-processing script, a listener or a transition.

 For example:

 import java.net.URLDecoder
 URLDecoder.decode(engineEnvironment.getParameter("locationCityStateZip"), "UTF-8")}