[feat] Allow for social authentication using Firebase
Not a complete implementation but functional. This addition will allow
you to sign up either with an email and password or will one of the
social media channels (Google, Facebook or GitHub). You can also login
with any of those methods. You will need to define the necessary .env
variables and setup an account in GitHub - and enable those channels.