[feat] Present a file upload button for inputType=upload
If you define an output parameter of inputType=upload then Leopard will
hide the input box and will replace any adject icons with a file upload
button. YOu can select a file and you will be given a progress as it
"uploads". The file is not really posted anywhere at the moment but that
will be configurable at some point.