:sparkles: feat: Assistant responses can have colored borders configured
If the output parameter "border" exists then it's value will be used as the left border  or the response or response chunks. If you want to control the borders for each response chunk the you will need a "borders" output parameter and an array of colors. borders= ["Red", "Green", "", "Blue"]